requirements.txt
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.2
SQLAlchemy==2.0.37
